+++ include/Ark/Builtins/Builtins.hpp
@@ -26,6 +26,7 @@ namespace Ark::internal::Builtins
     extern const Value falseSym;
     extern const Value trueSym;
     extern const Value nil;
+    extern const Value platform;
 
     ARK_API extern const std::vector<std::pair<std::string, Value>> builtins;
 
@@ -34,12 +35,12 @@ namespace Ark::internal::Builtins
     // ------------------------------
     namespace List
     {
-        Value reverseList(std::vector<Value>& n, VM* vm);  // list:reverse, single arg
-        Value findInList(std::vector<Value>& n, VM* vm);   // list:find, 2 arguments
-        Value sliceList(std::vector<Value>& n, VM* vm);    // list:slice, 4 arguments
-        Value sort_(std::vector<Value>& n, VM* vm);        // list:sort, 1 argument
-        Value fill(std::vector<Value>& n, VM* vm);         // list:fill, 2 arguments
-        Value setListAt(std::vector<Value>& n, VM* vm);    // list:setAt, 3 arguments
+        Value reverseList(std::vector<Value>& n, VM* vm);  // builtin__list:reverse, single arg
+        Value findInList(std::vector<Value>& n, VM* vm);   // builtin__list:find, 2 arguments
+        Value sliceList(std::vector<Value>& n, VM* vm);    // builtin__list:slice, 4 arguments
+        Value sort_(std::vector<Value>& n, VM* vm);        // builtin__list:sort, 1 argument
+        Value fill(std::vector<Value>& n, VM* vm);         // builtin__list:fill, 2 arguments
+        Value setListAt(std::vector<Value>& n, VM* vm);    // builtin__list:setAt, 3 arguments
     }
 
     namespace IO
@@ -47,14 +48,14 @@ namespace Ark::internal::Builtins
         Value print(std::vector<Value>& n, VM* vm);         // print, multiple arguments
         Value puts_(std::vector<Value>& n, VM* vm);         // puts, multiple arguments
         Value input(std::vector<Value>& n, VM* vm);         // input, 0 or 1 argument
-        Value writeFile(std::vector<Value>& n, VM* vm);     // io:writeFile, 2 arguments
-        Value appendToFile(std::vector<Value>& n, VM* vm);  // io:appendToFile, 2 arguments
-        Value readFile(std::vector<Value>& n, VM* vm);      // io:readFile, 1 argument
-        Value fileExists(std::vector<Value>& n, VM* vm);    // io:fileExists?, 1 argument
-        Value listFiles(std::vector<Value>& n, VM* vm);     // io:listFiles, 1 argument
-        Value isDirectory(std::vector<Value>& n, VM* vm);   // io:isDir?, 1 argument
-        Value makeDir(std::vector<Value>& n, VM* vm);       // io:makeDir, 1 argument
-        Value removeFiles(std::vector<Value>& n, VM* vm);   // io:removeFiles, multiple arguments
+        Value writeFile(std::vector<Value>& n, VM* vm);     // builtin__io:writeFile, 2 arguments
+        Value appendToFile(std::vector<Value>& n, VM* vm);  // builtin__io:appendToFile, 2 arguments
+        Value readFile(std::vector<Value>& n, VM* vm);      // builtin__io:readFile, 1 argument
+        Value fileExists(std::vector<Value>& n, VM* vm);    // builtin__io:fileExists?, 1 argument
+        Value listFiles(std::vector<Value>& n, VM* vm);     // builtin__io:listFiles, 1 argument
+        Value isDirectory(std::vector<Value>& n, VM* vm);   // builtin__io:isDir?, 1 argument
+        Value makeDir(std::vector<Value>& n, VM* vm);       // builtin__io:makeDir, 1 argument
+        Value removeFile(std::vector<Value>& n, VM* vm);    // builtin__io:removeFile, multiple arguments
     }
 
     namespace Time
@@ -64,30 +65,30 @@ namespace Ark::internal::Builtins
 
     namespace System
     {
-        Value system_(std::vector<Value>& n, VM* vm);  // sys:exec, 1 argument
-        Value sleep(std::vector<Value>& n, VM* vm);    // sleep, 1 argument
-        Value exit_(std::vector<Value>& n, VM* vm);    // sys:exit, 1 argument
+        Value system_(std::vector<Value>& n, VM* vm);  // builtin__sys:exec, 1 argument
+        Value sleep(std::vector<Value>& n, VM* vm);    // builtin__sys:sleep, 1 argument
+        Value exit_(std::vector<Value>& n, VM* vm);    // builtin__sys:exit, 1 argument
     }
 
     namespace String
     {
-        Value format(std::vector<Value>& n, VM* vm);       // string:format, multiple arguments
-        Value findSubStr(std::vector<Value>& n, VM* vm);   // string:find, 2 arguments
-        Value removeAtStr(std::vector<Value>& n, VM* vm);  // string:removeAt, 2 arguments
-        Value ord(std::vector<Value>& n, VM* vm);          // string:ord, 1 arguments
-        Value chr(std::vector<Value>& n, VM* vm);          // string:chr, 1 arguments
-        Value setStringAt(std::vector<Value>& n, VM* vm);  // string::setAt, 3 arguments
+        Value format(std::vector<Value>& n, VM* vm);       // format, multiple arguments
+        Value findSubStr(std::vector<Value>& n, VM* vm);   // builtin__string:find, 2 arguments
+        Value removeAtStr(std::vector<Value>& n, VM* vm);  // builtin__string:removeAt, 2 arguments
+        Value ord(std::vector<Value>& n, VM* vm);          // builtin__string:ord, 1 arguments
+        Value chr(std::vector<Value>& n, VM* vm);          // builtin__string:chr, 1 arguments
+        Value setStringAt(std::vector<Value>& n, VM* vm);  // builtin__string::setAt, 3 arguments
     }
 
     namespace Mathematics
     {
-        Value exponential(std::vector<Value>& n, VM* vm);  // math:exp, 1 argument
-        Value logarithm(std::vector<Value>& n, VM* vm);    // math:ln, 1 argument
-        Value ceil_(std::vector<Value>& n, VM* vm);        // math:ceil, 1 argument
-        Value floor_(std::vector<Value>& n, VM* vm);       // math:floor, 1 argument
-        Value round_(std::vector<Value>& n, VM* vm);       // math:round, 1 argument
-        Value isnan_(std::vector<Value>& n, VM* vm);       // math:NaN?, 1 argument
-        Value isinf_(std::vector<Value>& n, VM* vm);       // math:Inf?, 1 argument
+        Value exponential(std::vector<Value>& n, VM* vm);  // builtin__math:exp, 1 argument
+        Value logarithm(std::vector<Value>& n, VM* vm);    // builtin__math:ln, 1 argument
+        Value ceil_(std::vector<Value>& n, VM* vm);        // builtin__math:ceil, 1 argument
+        Value floor_(std::vector<Value>& n, VM* vm);       // builtin__math:floor, 1 argument
+        Value round_(std::vector<Value>& n, VM* vm);       // builtin__math:round, 1 argument
+        Value isnan_(std::vector<Value>& n, VM* vm);       // builtin__math:NaN?, 1 argument
+        Value isinf_(std::vector<Value>& n, VM* vm);       // builtin__math:Inf?, 1 argument
 
         extern const Value pi_;
         extern const Value e_;
@@ -95,18 +96,18 @@ namespace Ark::internal::Builtins
         extern const Value inf_;
         extern const Value nan_;
 
-        Value cos_(std::vector<Value>& n, VM* vm);    // math:cos, 1 argument
-        Value sin_(std::vector<Value>& n, VM* vm);    // math:sin, 1 argument
-        Value tan_(std::vector<Value>& n, VM* vm);    // math:tan, 1 argument
-        Value acos_(std::vector<Value>& n, VM* vm);   // math:arccos, 1 argument
-        Value asin_(std::vector<Value>& n, VM* vm);   // math:arcsin, 1 argument
-        Value atan_(std::vector<Value>& n, VM* vm);   // math:arctan, 1 argument
-        Value cosh_(std::vector<Value>& n, VM* vm);   // math:cosh, 1 argument
-        Value sinh_(std::vector<Value>& n, VM* vm);   // math:sinh, 1 argument
-        Value tanh_(std::vector<Value>& n, VM* vm);   // math:tanh, 1 argument
-        Value acosh_(std::vector<Value>& n, VM* vm);  // math:acosh, 1 argument
-        Value asinh_(std::vector<Value>& n, VM* vm);  // math:asinh, 1 argument
-        Value atanh_(std::vector<Value>& n, VM* vm);  // math:atanh, 1 argument
+        Value cos_(std::vector<Value>& n, VM* vm);    // builtin__math:cos, 1 argument
+        Value sin_(std::vector<Value>& n, VM* vm);    // builtin__math:sin, 1 argument
+        Value tan_(std::vector<Value>& n, VM* vm);    // builtin__math:tan, 1 argument
+        Value acos_(std::vector<Value>& n, VM* vm);   // builtin__math:arccos, 1 argument
+        Value asin_(std::vector<Value>& n, VM* vm);   // builtin__math:arcsin, 1 argument
+        Value atan_(std::vector<Value>& n, VM* vm);   // builtin__math:arctan, 1 argument
+        Value cosh_(std::vector<Value>& n, VM* vm);   // builtin__math:cosh, 1 argument
+        Value sinh_(std::vector<Value>& n, VM* vm);   // builtin__math:sinh, 1 argument
+        Value tanh_(std::vector<Value>& n, VM* vm);   // builtin__math:tanh, 1 argument
+        Value acosh_(std::vector<Value>& n, VM* vm);  // builtin__math:acosh, 1 argument
+        Value asinh_(std::vector<Value>& n, VM* vm);  // builtin__math:asinh, 1 argument
+        Value atanh_(std::vector<Value>& n, VM* vm);  // builtin__math:atanh, 1 argument
 
         Value random(std::vector<Value>& n, VM* vm);  // random, 0-2 args
     }