+++ src/arkreactor/Builtins/System.cpp
@@ -30,16 +30,6 @@ namespace Ark::internal::Builtins::System
         };
     }
 
-    /**
-     * @name sys:exec
-     * @brief Execute a system specific command
-     * @details Return the output of the command as a String, or nil if it was disabled in the ArkScript build
-     * @param command the command to execute, as a String
-     * =begin
-     * (sys:exec "echo hello")
-     * =end
-     * @author https://github.com/SuperFola
-     */
     // cppcheck-suppress constParameterReference
     Value system_(std::vector<Value>& n, VM* vm [[maybe_unused]])
     {
@@ -63,16 +53,6 @@ namespace Ark::internal::Builtins::System
 #endif  // ARK_ENABLE_SYSTEM
     }
 
-    /**
-     * @name sys:sleep
-     * @brief Sleep for a given duration (in milliseconds)
-     * @details Return nil
-     * @param duration a Number representing a duration
-     * =begin
-     * (sys:sleep 1000)  # sleep for 1 second
-     * =end
-     * @author https://github.com/SuperFola
-     */
     // cppcheck-suppress constParameterReference
     Value sleep(std::vector<Value>& n, VM* vm [[maybe_unused]])
     {
@@ -88,16 +68,6 @@ namespace Ark::internal::Builtins::System
         return nil;
     }
 
-    /**
-     * @name sys:exit
-     * @brief Reverse a given list and return a new one
-     * @details Any code after this function call won't be executed
-     * @param exitCode usually 0 for success and 1 for errors
-     * =begin
-     * (sys:exit 0)  # halt the virtual machine with given exit code (success)
-     * =end
-     * @author https://github.com/SuperFola
-     */
     // cppcheck-suppress constParameterReference
     Value exit_(std::vector<Value>& n, VM* vm)
     {